requirements.txt
tensorflow ~= 2.13
flask ~= 3.0
pillow ~= 10.4
numpy ~= 1.24